+++ packages/jest-core/src/TestScheduler.ts
@@ -44,6 +44,7 @@ export type TestSchedulerContext = {
   firstRun: boolean;
   previousSuccess: boolean;
   changedFiles?: Set<Config.Path>;
+  sourcesRelatedToTestsInChangedFiles?: Set<Config.Path>;
 };
 export default class TestScheduler {
   private _dispatcher: ReporterDispatcher;
@@ -175,12 +176,14 @@ export default class TestScheduler {
       showStatus: !runInBand,
     });
 
-    const testRunners = Object.create(null);
+    const testRunners: {[key: string]: TestRunner} = Object.create(null);
     contexts.forEach(({config}) => {
       if (!testRunners[config.runner]) {
         const Runner: typeof TestRunner = require(config.runner);
         testRunners[config.runner] = new Runner(this._globalConfig, {
-          changedFiles: this._context && this._context.changedFiles,
+          changedFiles: this._context?.changedFiles,
+          sourcesRelatedToTestsInChangedFiles: this._context
+            ?.sourcesRelatedToTestsInChangedFiles,
         });
       }
     });
@@ -272,7 +275,9 @@ export default class TestScheduler {
     if (!isDefault && collectCoverage) {
       this.addReporter(
         new CoverageReporter(this._globalConfig, {
-          changedFiles: this._context && this._context.changedFiles,
+          changedFiles: this._context?.changedFiles,
+          sourcesRelatedToTestsInChangedFiles: this._context
+            ?.sourcesRelatedToTestsInChangedFiles,
         }),
       );
     }
@@ -302,7 +307,9 @@ export default class TestScheduler {
     if (collectCoverage) {
       this.addReporter(
         new CoverageReporter(this._globalConfig, {
-          changedFiles: this._context && this._context.changedFiles,
+          changedFiles: this._context?.changedFiles,
+          sourcesRelatedToTestsInChangedFiles: this._context
+            ?.sourcesRelatedToTestsInChangedFiles,
         }),
       );
     }