+++ packages/jest-runner/src/types.ts
@@ -51,6 +51,7 @@ export type TestRunnerOptions = {
 
 export type TestRunnerContext = {
   changedFiles?: Set<Config.Path>;
+  sourcesRelatedToTestsInChangedFiles?: Set<Config.Path>;
 };
 
 export type TestRunnerSerializedContext = {
@@ -63,7 +64,6 @@ export type WatcherState = {
 };
 export interface TestWatcher extends EventEmitter {
   state: WatcherState;
-  new ({isWatchMode}: {isWatchMode: boolean}): TestWatcher;
   setState(state: WatcherState): void;
   isInterrupted(): boolean;
   isWatchMode(): boolean;