+++ packages/jest-runtime/src/index.ts
@@ -5,9 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {URL, fileURLToPath} from 'url';
+import {URL, fileURLToPath, pathToFileURL} from 'url';
 import * as path from 'path';
-import {Script, compileFunction} from 'vm';
+import {
+  Script,
+  // @ts-ignore: experimental, not added to the types
+  SourceTextModule,
+  // @ts-ignore: experimental, not added to the types
+  SyntheticModule,
+  Context as VMContext,
+  // @ts-ignore: experimental, not added to the types
+  Module as VMModule,
+  compileFunction,
+} from 'vm';
 import * as nativeModule from 'module';
 import type {Config, Global} from '@jest/types';
 import type {
@@ -76,6 +86,8 @@ type ResolveOptions = Parameters<typeof require.resolve>[1];
 type BooleanObject = Record<string, boolean>;
 type CacheFS = {[path: string]: string};
 
+type RequireCache = {[key: string]: Module};
+
 namespace Runtime {
   export type Context = JestContext;
   // ditch this export when moving to esm - for now we need it for to avoid faulty type elision
@@ -106,6 +118,8 @@ const EVAL_RESULT_VARIABLE = 'Object.<anonymous>';
 
 type RunScriptEvalResult = {[EVAL_RESULT_VARIABLE]: ModuleWrapper};
 
+const runtimeSupportsVmModules = typeof SyntheticModule === 'function';
+
 /* eslint-disable-next-line no-redeclare */
 class Runtime {
   private _cacheFS: CacheFS;
@@ -125,7 +139,7 @@ class Runtime {
   private _moduleMocker: typeof jestMock;
   private _isolatedModuleRegistry: ModuleRegistry | null;
   private _moduleRegistry: ModuleRegistry;
-  private _needsCoverageMapped: Set<string>;
+  private _esmoduleRegistry: Map<string, Promise<VMModule>>;
   private _resolver: Resolver;
   private _shouldAutoMock: boolean;
   private _shouldMockModuleCache: BooleanObject;
@@ -140,6 +154,7 @@ class Runtime {
   private _virtualMocks: BooleanObject;
   private _moduleImplementation?: typeof nativeModule.Module;
   private jestObjectCaches: Map<string, Jest>;
+  private _hasWarnedAboutRequireCacheModification = false;
 
   constructor(
     config: Config.ProjectConfig,
@@ -156,6 +171,7 @@ class Runtime {
       collectCoverageFrom: [],
       collectCoverageOnlyFrom: undefined,
       coverageProvider: 'babel',
+      sourcesRelatedToTestsInChangedFiles: undefined,
     };
     this._currentlyExecutingModulePath = '';
     this._environment = environment;
@@ -169,7 +185,7 @@ class Runtime {
     this._isolatedModuleRegistry = null;
     this._isolatedMockRegistry = null;
     this._moduleRegistry = new Map();
-    this._needsCoverageMapped = new Set();
+    this._esmoduleRegistry = new Map();
     this._resolver = resolver;
     this._scriptTransformer = new ScriptTransformer(config);
     this._shouldAutoMock = config.automock;
@@ -302,6 +318,140 @@ class Runtime {
     return cliOptions;
   }
 
+  // unstable as it should be replaced by https://github.com/nodejs/modules/issues/393, and we don't want people to use it
+  unstable_shouldLoadAsEsm = Resolver.unstable_shouldLoadAsEsm;
+
+  private async loadEsmModule(
+    modulePath: Config.Path,
+    query = '',
+  ): Promise<VMModule> {
+    const cacheKey = modulePath + query;
+
+    if (!this._esmoduleRegistry.has(cacheKey)) {
+      invariant(
+        typeof this._environment.getVmContext === 'function',
+        'ES Modules are only supported if your test environment has the `getVmContext` function',
+      );
+
+      const context = this._environment.getVmContext();
+
+      invariant(context);
+
+      if (this._resolver.isCoreModule(modulePath)) {
+        const core = await this._importCoreModule(modulePath, context);
+        this._esmoduleRegistry.set(cacheKey, core);
+        return core;
+      }
+
+      const transformedFile = this.transformFile(modulePath, {
+        isInternalModule: false,
+        supportsDynamicImport: true,
+        supportsStaticESM: true,
+      });
+
+      const module = new SourceTextModule(transformedFile.code, {
+        context,
+        identifier: modulePath,
+        importModuleDynamically: this.linkModules.bind(this),
+        initializeImportMeta(meta: ImportMeta) {
+          meta.url = pathToFileURL(modulePath).href;
+        },
+      });
+
+      this._esmoduleRegistry.set(
+        cacheKey,
+        // we wanna put the linking promise in the cache so modules loaded in
+        // parallel can all await it. We then await it synchronously below, so
+        // we shouldn't get any unhandled rejections
+        module
+          .link(this.linkModules.bind(this))
+          .then(() => module.evaluate())
+          .then(() => module),
+      );
+    }
+
+    const module = this._esmoduleRegistry.get(cacheKey);
+
+    invariant(module);
+
+    return module;
+  }
+
+  private linkModules(specifier: string, referencingModule: VMModule) {
+    if (specifier === '@jest/globals') {
+      const fromCache = this._esmoduleRegistry.get('@jest/globals');
+
+      if (fromCache) {
+        return fromCache;
+      }
+      const globals = this.getGlobalsForEsm(
+        referencingModule.identifier,
+        referencingModule.context,
+      );
+      this._esmoduleRegistry.set('@jest/globals', globals);
+
+      return globals;
+    }
+
+    const resolved = this._resolveModule(
+      referencingModule.identifier,
+      specifier,
+    );
+
+    if (
+      this._resolver.isCoreModule(resolved) ||
+      this.unstable_shouldLoadAsEsm(resolved)
+    ) {
+      return this.loadEsmModule(resolved);
+    }
+
+    return this.loadCjsAsEsm(
+      referencingModule.identifier,
+      resolved,
+      referencingModule.context,
+    );
+  }
+
+  async unstable_importModule(
+    from: Config.Path,
+    moduleName?: string,
+  ): Promise<void> {
+    invariant(
+      runtimeSupportsVmModules,
+      'You need to run with a version of node that supports ES Modules in the VM API.',
+    );
+
+    const modulePath = this._resolveModule(from, moduleName);
+
+    return this.loadEsmModule(modulePath);
+  }
+
+  private async loadCjsAsEsm(
+    from: Config.Path,
+    modulePath: Config.Path,
+    context: VMContext,
+  ) {
+    // CJS loaded via `import` should share cache with other CJS: https://github.com/nodejs/modules/issues/503
+    const cjs = this.requireModuleOrMock(from, modulePath);
+
+    const module = new SyntheticModule(
+      ['default'],
+      function () {
+        // @ts-ignore: TS doesn't know what `this` is
+        this.setExport('default', cjs);
+      },
+      {context, identifier: modulePath},
+    );
+
+    await module.link(() => {
+      throw new Error('This should never happen');
+    });
+
+    await module.evaluate();
+
+    return module;
+  }
+
   requireModule<T = unknown>(
     from: Config.Path,
     moduleName?: string,
@@ -329,11 +479,6 @@ class Runtime {
       this._explicitShouldMock[moduleID] !== false
     ) {
       modulePath = manualMock;
-    }
-
-    if (moduleName === '@jest/globals') {
-      // @ts-ignore: we don't care that it's not assignable to T
-      return this.getGlobalsForFile(from);
     }
 
     if (moduleName && this._resolver.isCoreModule(moduleName)) {
@@ -526,12 +671,18 @@ class Runtime {
     };
   }
 
-  requireModuleOrMock(from: Config.Path, moduleName: string): unknown {
+  requireModuleOrMock<T = unknown>(from: Config.Path, moduleName: string): T {
+    // this module is unmockable
+    if (moduleName === '@jest/globals') {
+      // @ts-ignore: we don't care that it's not assignable to T
+      return this.getGlobalsForCjs(from);
+    }
+
     try {
       if (this._shouldMock(from, moduleName)) {
-        return this.requireMock(from, moduleName);
+        return this.requireMock<T>(from, moduleName);
       } else {
-        return this.requireModule(from, moduleName);
+        return this.requireModule<T>(from, moduleName);
       }
     } catch (e) {
       const moduleNotFound = Resolver.tryCastModuleNotFoundError(e);
@@ -577,6 +728,7 @@ class Runtime {
     this._isolatedMockRegistry = null;
     this._mockRegistry.clear();
     this._moduleRegistry.clear();
+    this._esmoduleRegistry.clear();
 
     if (this._environment) {
       if (this._environment.global) {
@@ -641,20 +793,9 @@ class Runtime {
       });
   }
 
-  getSourceMapInfo(coveredFiles: Set<string>): Record<string, string> {
-    return Object.keys(this._sourceMapRegistry).reduce<Record<string, string>>(
-      (result, sourcePath) => {
-        if (
-          coveredFiles.has(sourcePath) &&
-          this._needsCoverageMapped.has(sourcePath) &&
-          fs.existsSync(this._sourceMapRegistry[sourcePath])
-        ) {
-          result[sourcePath] = this._sourceMapRegistry[sourcePath];
-        }
-        return result;
-      },
-      {},
-    );
+  // TODO - remove in Jest 26
+  getSourceMapInfo(_coveredFiles: Set<string>): Record<string, string> {
+    return {};
   }
 
   getSourceMaps(): SourceMapRegistry {
@@ -795,23 +936,8 @@ class Runtime {
     Object.defineProperty(localModule, 'require', {
       value: this._createRequireImplementation(localModule, options),
     });
-    const transformedFile = this._scriptTransformer.transform(
-      filename,
-      this._getFullTransformationOptions(options),
-      this._cacheFS[filename],
-    );
 
-    // we only care about non-internal modules
-    if (!options || !options.isInternalModule) {
-      this._fileTransforms.set(filename, transformedFile);
-    }
-
-    if (transformedFile.sourceMapPath) {
-      this._sourceMapRegistry[filename] = transformedFile.sourceMapPath;
-      if (transformedFile.mapCoverage) {
-        this._needsCoverageMapped.add(filename);
-      }
-    }
+    const transformedFile = this.transformFile(filename, options);
 
     let compiledFunction: ModuleWrapper | null = null;
 
@@ -872,10 +998,7 @@ class Runtime {
       return;
     }
 
-    const jestObject = this._createJestObjectFor(
-      filename,
-      localModule.require as LocalModuleRequire,
-    );
+    const jestObject = this._createJestObjectFor(filename);
 
     this.jestObjectCaches.set(filename, jestObject);
 
@@ -907,6 +1030,24 @@ class Runtime {
     this._currentlyExecutingModulePath = lastExecutingModulePath;
   }
 
+  private transformFile(filename: string, options?: InternalModuleOptions) {
+    const transformedFile = this._scriptTransformer.transform(
+      filename,
+      this._getFullTransformationOptions(options),
+      this._cacheFS[filename],
+    );
+
+    // we only care about non-internal modules
+    if (!options || !options.isInternalModule) {
+      this._fileTransforms.set(filename, transformedFile);
+    }
+
+    if (transformedFile.sourceMapPath) {
+      this._sourceMapRegistry[filename] = transformedFile.sourceMapPath;
+    }
+    return transformedFile;
+  }
+
   private createScriptFromCode(scriptSource: string, filename: string) {
     try {
       return new Script(this.wrapCodeInModuleWrapper(scriptSource), {
@@ -930,6 +1071,24 @@ class Runtime {
     }
 
     return require(moduleName);
+  }
+
+  private _importCoreModule(moduleName: string, context: VMContext) {
+    const required = this._requireCoreModule(moduleName);
+
+    return new SyntheticModule(
+      ['default', ...Object.keys(required)],
+      function () {
+        // @ts-ignore: TS doesn't know what `this` is
+        this.setExport('default', required);
+        Object.entries(required).forEach(([key, value]) => {
+          // @ts-ignore: TS doesn't know what `this` is
+          this.setExport(key, value);
+        });
+      },
+      // should identifier be `node://${moduleName}`?
+      {context, identifier: moduleName},
+    );
   }
 
   private _getMockedNativeModule(): typeof nativeModule.Module {
@@ -966,13 +1125,20 @@ class Runtime {
     // should we implement the class ourselves?
     class Module extends nativeModule.Module {}
 
+    Object.entries(nativeModule.Module).forEach(([key, value]) => {
+      // @ts-ignore
+      Module[key] = value;
+    });
+
     Module.Module = Module;
 
     if ('createRequire' in nativeModule) {
       Module.createRequire = createRequire;
     }
     if ('createRequireFromPath' in nativeModule) {
-      Module.createRequireFromPath = (filename: string | URL) => {
+      Module.createRequireFromPath = function createRequireFromPath(
+        filename: string | URL,
+      ) {
         if (typeof filename !== 'string') {
           const error = new TypeError(
             `The argument 'filename' must be string. Received '${filename}'.${
@@ -989,7 +1155,7 @@ class Runtime {
       };
     }
     if ('syncBuiltinESMExports' in nativeModule) {
-      Module.syncBuiltinESMExports = () => {};
+      Module.syncBuiltinESMExports = function syncBuiltinESMExports() {};
     }
 
     this._moduleImplementation = Module;
@@ -1116,11 +1282,38 @@ class Runtime {
           this,
           from.filename,
         )) as LocalModuleRequire;
-    moduleRequire.cache = Object.create(null);
     moduleRequire.extensions = Object.create(null);
     moduleRequire.requireActual = this.requireActual.bind(this, from.filename);
     moduleRequire.requireMock = this.requireMock.bind(this, from.filename);
     moduleRequire.resolve = resolve;
+    moduleRequire.cache = (() => {
+      const notPermittedMethod = () => {
+        if (!this._hasWarnedAboutRequireCacheModification) {
+          this._environment.global.console.warn(
+            '`require.cache` modification is not permitted',
+          );
+
+          this._hasWarnedAboutRequireCacheModification = true;
+        }
+        return true;
+      };
+      return new Proxy<RequireCache>(Object.create(null), {
+        defineProperty: notPermittedMethod,
+        deleteProperty: notPermittedMethod,
+        get: (_target, key) =>
+          typeof key === 'string' ? this._moduleRegistry.get(key) : undefined,
+        getOwnPropertyDescriptor() {
+          return {
+            configurable: true,
+            enumerable: true,
+          };
+        },
+        has: (_target, key) =>
+          typeof key === 'string' && this._moduleRegistry.has(key),
+        ownKeys: () => Array.from(this._moduleRegistry.keys()),
+        set: notPermittedMethod,
+      });
+    })();
 
     Object.defineProperty(moduleRequire, 'main', {
       enumerable: true,
@@ -1139,10 +1332,7 @@ class Runtime {
     return moduleRequire;
   }
 
-  private _createJestObjectFor(
-    from: Config.Path,
-    localRequire: LocalModuleRequire,
-  ): Jest {
+  private _createJestObjectFor(from: Config.Path): Jest {
     const disableAutomock = () => {
       this._shouldAutoMock = false;
       return jestObject;
@@ -1273,8 +1463,8 @@ class Runtime {
       isMockFunction: this._moduleMocker.isMockFunction,
       isolateModules,
       mock,
-      requireActual: localRequire.requireActual,
-      requireMock: localRequire.requireMock,
+      requireActual: this.requireActual.bind(this, from),
+      requireMock: this.requireMock.bind(this, from),
       resetAllMocks,
       resetModuleRegistry: resetModules,
       resetModules,
@@ -1355,11 +1545,52 @@ class Runtime {
     throw e;
   }
 
-  private getGlobalsForFile(from: Config.Path): JestGlobalsValues {
+  private getGlobalsForCjs(from: Config.Path): JestGlobalsValues {
     const jest = this.jestObjectCaches.get(from);
 
     invariant(jest, 'There should always be a Jest object already');
 
+    return {...this.getGlobalsFromEnvironment(), jest};
+  }
+
+  private async getGlobalsForEsm(
+    from: Config.Path,
+    context: VMContext,
+  ): Promise<VMModule> {
+    let jest = this.jestObjectCaches.get(from);
+
+    if (!jest) {
+      jest = this._createJestObjectFor(from);
+
+      this.jestObjectCaches.set(from, jest);
+    }
+
+    const globals: JestGlobalsValues = {
+      ...this.getGlobalsFromEnvironment(),
+      jest,
+    };
+
+    const module = new SyntheticModule(
+      Object.keys(globals),
+      function () {
+        Object.entries(globals).forEach(([key, value]) => {
+          // @ts-ignore: TS doesn't know what `this` is
+          this.setExport(key, value);
+        });
+      },
+      {context, identifier: '@jest/globals'},
+    );
+
+    await module.link(() => {
+      throw new Error('This should never happen');
+    });
+
+    await module.evaluate();
+
+    return module;
+  }
+
+  private getGlobalsFromEnvironment(): Omit<JestGlobalsValues, 'jest'> {
     return {
       afterAll: this._environment.global.afterAll,
       afterEach: this._environment.global.afterEach,
@@ -1370,7 +1601,6 @@ class Runtime {
       fdescribe: this._environment.global.fdescribe,
       fit: this._environment.global.fit,
       it: this._environment.global.it,
-      jest,
       test: this._environment.global.test,
       xdescribe: this._environment.global.xdescribe,
       xit: this._environment.global.xit,